new meeting screen bootstrapped
