+++ src/types/navigators/meeting.ts
@@ -1,0 +1,35 @@
+import { RouteProp } from "@react-navigation/native";
+import { NativeStackNavigationProp, NativeStackScreenProps } from "@react-navigation/native-stack";
+import { MeetingTime } from "../../generated/client";
+
+/**
+ * Fund selection navigator
+ */
+namespace MeetingNavigator {
+
+  /**
+   * Routes
+   */
+  export type Routes = {
+    meetingTimes: undefined;
+    newMeeting: { meetingTime: MeetingTime };
+  };
+
+  /**
+   * Navigation properties
+   */
+  export type NavigationProps<RouteName extends keyof Routes = keyof Routes> = NativeStackNavigationProp<Routes, RouteName>;
+
+  /**
+   * Route properties
+   */
+  export type RouteProps<RouteName extends keyof Routes = keyof Routes> = RouteProp<Routes, RouteName>;
+
+  /**
+   * Screen properties
+   */
+  export type ScreenProps<RouteName extends keyof Routes = keyof Routes> = NativeStackScreenProps<Routes, RouteName>;
+
+}
+
+export default MeetingNavigator;
